tests/e2e-test/requirements.txt
pytest-playwright
pytest-reporter-html1
python-dotenv
pytest-check
pytest-html
py
beautifulsoup4
Pillow